feat: add client to kvg-collector
