+++ packages/kvg-collector/main.go
@@ -1,149 +1,44 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"net/url"
-	"strings"
+	"time"
 
-	funk "github.com/thoas/go-funk"
+	"github.com/go-co-op/gocron"
+	"github.com/kiel-live/kiel-live/packages/client"
+	protocol "github.com/kiel-live/kiel-live/packages/pub-sub-proto"
 )
 
-var baseURL = "https://www.kvg-kiel.de"
-var stopURL = baseURL + "/internetservice/services/passageInfo/stopPassages/stop"
-var tripURL = baseURL + "/internetservice/services/tripInfo/tripPassages"
-var vehiclesURL = baseURL + "/internetservice/geoserviceDispatcher/services/vehicleinfo/vehicles"
-var stopsURL = baseURL + "/internetservice/geoserviceDispatcher/services/stopinfo/stops"
-
 func main() {
-	stops := getStops()
-	fmt.Println(stops.Stops[0])
+	client := client.NewWebSocketClient("localhost:4000", func(msg protocol.ClientMessage) {
+		fmt.Println("huhu > " + msg.Data())
+	})
 
-	vehicles := getVehicles()
-	fmt.Println(vehicles.Vehicles[0])
+	client.Connect()
+	defer client.Disconnect()
 
-	stop := getStop(stops.Stops[0].ShortName)
-	fmt.Println(stop)
+	client.Subscribe(protocol.SubscribedChannelName)
 
-	trip := getTrip(stop.Departures[0].TripID)
-	fmt.Println(trip)
+	s := gocron.NewScheduler(time.UTC)
+	s.SetMaxConcurrentJobs(1, gocron.RescheduleMode) // prevent parallel execution and skip if last run hasn't finished yet
+	s.Every(30).Seconds().Do(func() {
+		fetchData()
+	})
+
+	s.StartBlocking()
 }
 
-func post(url string, data url.Values) ([]byte, error) {
-	resp, err := http.Post(url, "application/x-www-form-urlencoded", strings.NewReader(data.Encode()))
-	if err != nil {
-		println(err)
-		return nil, err
-	}
-	defer resp.Body.Close()
-	fmt.Println(resp.Status)
-	body, err := io.ReadAll(resp.Body)
-	return body, err
-}
+func fetchData() {
+	fmt.Println("now")
+	// stops := getStops()
+	// fmt.Println(stops.Stops[0])
 
-type stop struct {
-	ID        string `json:"id"`
-	ShortName string `json:"shortName"`
-	Name      string `json:"name"`
-	Latitude  int    `json:"latitude"`
-	Longitude int    `json:"longitude"`
-}
+	// vehicles := getVehicles()
+	// fmt.Println(vehicles.Vehicles[0])
 
-type stops struct {
-	Stops []stop `json:"stops"`
-}
+	// stop := getStop(stops.Stops[0].ShortName)
+	// fmt.Println(stop)
 
-func getStops() stops {
-	data := url.Values{}
-	data.Set("top", "324000000")
-	data.Set("bottom", "-324000000")
-	data.Set("left", "-648000000")
-	data.Set("right", "648000000")
-
-	body, _ := post(stopsURL, data)
-	fmt.Printf("Body: %s\n", body)
-	var stops stops
-	if err := json.Unmarshal(body, &stops); err != nil {
-		log.Fatalf("Parse response failed, reason: %v \n", err)
-	}
-	return stops
-}
-
-type vehicle struct {
-	ID        string `json:"id"`
-	Name      string `json:"name"`
-	Heading   int    `json:"heading"`
-	Latitude  int    `json:"latitude"`
-	Longitude int    `json:"longitude"`
-	TripID    string `json:"tripId"`
-}
-
-type vehicles struct {
-	Vehicles []vehicle `json:"vehicles"`
-}
-
-func getVehicles() vehicles {
-	body, _ := post(vehiclesURL, nil)
-	var vehicles vehicles
-	if err := json.Unmarshal(body, &vehicles); err != nil {
-		log.Fatalf("Parse response failed, reason: %v \n", err)
-	}
-
-	vehicles.Vehicles = funk.Filter(vehicles.Vehicles, func(vehicle vehicle) bool {
-		return vehicle.Latitude != 0
-	}).([]vehicle)
-
-	return vehicles
-}
-
-type departure struct {
-	TripID             string `json:"tripId"`
-	Status             string `json:"status"`
-	Stop               string `json:"plannedTime"`
-	ActualTime         string `json:"actualTime"`
-	ActualRelativeTime int    `json:"actualRelativeTime"`
-}
-
-type stopDepartures struct {
-	Departures []departure `json:"actual"`
-}
-
-func getStop(stopShortName string) stopDepartures {
-	data := url.Values{}
-	data.Set("stop", stopShortName)
-
-	resp, _ := post(stopURL, data)
-	var stop stopDepartures
-	if err := json.Unmarshal(resp, &stop); err != nil {
-		log.Fatalf("Parse response failed, reason: %v \n", err)
-	}
-	return stop
-}
-
-type tripStops struct {
-	Stop       stop   `json:"stop"`
-	Status     string `json:"status"`
-	ActualTime string `json:"actualTime"`
-}
-
-type trip struct {
-	Stops         []tripStops `json:"actual"`
-	OldStops      []tripStops `json:"old"`
-	DirectionText string      `json:"directionText"`
-	RouteName     string      `json:"routeName"`
-}
-
-func getTrip(tripID string) trip {
-	data := url.Values{}
-	data.Set("tripId", tripID)
-
-	resp, _ := post(tripURL, data)
-	var trip trip
-	if err := json.Unmarshal(resp, &trip); err != nil {
-		log.Fatalf("Parse response failed, reason: %v \n", err)
-	}
-	return trip
+	// trip := getTrip(stop.Departures[0].TripID)
+	// fmt.Println(trip)
 }
